refactor(infra): move the tcgen05 descriptor folds next to their C encoders
The compile-time folds of the tcgen05 matrix and instruction descriptors landed
in `tile_primitive/gemm_async/tcgen05.py`, one package away from the runtime C
encoders and the `codegen/header.py` unions whose bit layouts they mirror.
`cpp/descriptors.py` already states that charter and already owns all three
runtime encoders, so the folds belong there; the only tvm caller imports the
dense one back.

Moving them also lets the dense and block-scaled folds share what they always
shared: the transpose-dtype check and the nine descriptor bits that are
identical between them, which previously existed as two hand-kept copies. The
dense fold picks up the block-scaled fold's device parity test over those bits
as a result.

No behavior change: every encoder produces the same value it did before, and
the three are now public since a kernel outside this repo consumes two of them.
